requirements.txt
llama-index
openai==1.0.0
python-dotenv==0.19.0
chromadb
pypdf
phidata
